Move the pure-Python deblending reference into its own module
