+++ photutils/segmentation/detect.py
@@ -242,72 +242,6 @@ def _detect_sources(data, threshold, n_pixels, footprint, inverse_mask):
                                         slices=segm_slices)
 
 
-def _detect_sources_deblend(data, threshold, n_pixels, *, footprint,
-                            segment_mask):
-    """
-    Detect sources for a single multithreshold level during deblending.
-
-    This is the deblending analogue of `_detect_sources`. It differs in
-    that the detected segments keep their (possibly non-consecutive)
-    label numbers from `~scipy.ndimage.label`, the small segments are
-    removed with a bincount-based area filter (the per-label cutout
-    loop used by `_detect_sources` has a fixed per-label overhead that
-    dominates for the small cutouts and the many calls made during
-    deblending), and `None` is returned when fewer than two segments are
-    found.
-
-    Parameters
-    ----------
-    data : 2D `~numpy.ndarray`
-        The cutout data array for a single source.
-
-    threshold : float
-        The data value to be used for the detection threshold.
-
-    n_pixels : int
-        The minimum number of connected pixels, each greater than
-        ``threshold``, that an object must have to be detected.
-
-    footprint : array_like
-        A footprint that defines feature connections.
-
-    segment_mask : 2D bool `~numpy.ndarray`
-        A boolean mask of the source segment, with the same shape as
-        ``data``. Pixels outside the segment will not be included in
-        any source.
-
-    Returns
-    -------
-    segment_img : 2D int `~numpy.ndarray` or `None`
-        A 2D segmentation image, with the same shape as ``data``,
-        where sources are marked by different positive integer
-        values. A value of zero is reserved for the background. If
-        fewer than two sources are found then `None` is returned.
-    """
-    # NaN values compare as False, so NaN pixels are never included
-    # in any source. The comparison is never empty because the
-    # deblending thresholds are strictly below the source maximum.
-    segment_img = data > threshold
-    segment_img &= segment_mask
-
-    segment_img, n_labels = ndi_label(segment_img, structure=footprint)
-
-    # Remove objects with less than n_pixels
-    areas = np.bincount(segment_img.ravel())
-    keep = areas >= n_pixels
-    keep[0] = False
-    n_keep = np.count_nonzero(keep)
-    if n_keep <= 1:
-        return None
-
-    if n_keep < n_labels:
-        label_map = np.where(
-            keep, np.arange(areas.size, dtype=segment_img.dtype), 0)
-        segment_img = label_map[segment_img]
-
-    return segment_img
-
-
 @deprecated_renamed_argument('npixels', 'n_pixels', '3.0', until='4.0')
 def detect_sources(data, threshold, n_pixels, *, connectivity=8, mask=None):
     """
